First working solution with redirection refactor
